requirements.txt
pyttsx3
python-dotenv
opencv-python
dlib
numpy
Pillow
openpyxl
pandas
xlrd